Cleanup in new gconf adoption
- enhance error messages
- move message code to `msg.go`
- move handler code to `handlers.go`
- register configuration handler together with other message handlers
